EditRule and Rule List bug fixes
Changed the way we handle validating the rule in the edit rules dialog
to be more reliable.  Also report errors that happen when you run the
rule.

Minor updates to the rules list to includ erule type in the custom list
and to show "gear" behind custom rules in the main lists (sell/keep)