Fix the background of item list, edit and dropdowns
Blizzard changed the constants to a color object, we need to call
GetRGB() first, or use r/g/b directly.  Fix all the places where
we use background to handle it.

NOTE: this should move to a utility mixin so that we can just invoke
it all of the places.